+++ src/test/scala/au/com/agl/arc/extract/BytesExtractSuite.scala
@@ -28,6 +28,9 @@ class BytesExtractSuite extends FunSuite with BeforeAndAfter {
   val pathView = "pathView"
   val outputView = "outputView"
   val targetFile = getClass.getResource("/notes.xml.zip").toString
+  val emptyDirectory = FileUtils.getTempDirectoryPath() + "missing.binary" 
+  val missingDirectory = FileUtils.getTempDirectoryPath() + "/missing/missing.binary" 
+  val emptyWildcardDirectory = FileUtils.getTempDirectoryPath() + "*.binary" 
 
   before {
     implicit val spark = SparkSession
@@ -59,6 +62,7 @@ class BytesExtractSuite extends FunSuite with BeforeAndAfter {
       BytesExtract(
         name="dataset",
         description=None,
+        cols=Right(Nil),
         outputView=outputView, 
         input=Right(targetFile),
         authentication=None,
@@ -86,6 +90,7 @@ class BytesExtractSuite extends FunSuite with BeforeAndAfter {
       BytesExtract(
         name="dataset",
         description=None,
+        cols=Right(Nil),
         outputView=outputView, 
         input=Left(pathView),
         authentication=None,
@@ -98,5 +103,100 @@ class BytesExtractSuite extends FunSuite with BeforeAndAfter {
 
     assert(extractDataset.count == 2)
   }    
+
+  test("BytesExtract Empty Dataset") {
+    implicit val spark = session
+    import spark.implicits._
+    implicit val logger = LoggerFactory.getLogger(spark.sparkContext.applicationId)
+
+    val cols = 
+      BinaryColumn(
+        id="1",
+        name="binaryDatum",
+        description=None,
+        nullable=true,
+        nullReplacementValue=None,
+        trim=false,
+        nullableValues=Nil, 
+        encoding=EncodingTypeBase64,
+        metadata=None
+      ) :: Nil    
+
+    // try with wildcard
+    val thrown0 = intercept[Exception with DetailException] {
+      val extractDataset = extract.BytesExtract.extract(
+        BytesExtract(
+          name="dataset",
+          description=None,
+          cols=Right(Nil),
+          outputView=outputView, 
+          input=Right(emptyWildcardDirectory),
+          authentication=None,
+          persist=false,
+          numPartitions=None,
+          contiguousIndex=true,
+          params=Map.empty
+        )        
+      )
+    }
+    assert(thrown0.getMessage === "BytesExtract has produced 0 columns and no schema has been provided to create an empty dataframe.")
+    
+    // try without providing column metadata
+    val thrown1 = intercept[Exception with DetailException] {
+      val extractDataset = extract.BytesExtract.extract(
+        BytesExtract(
+          name="dataset",
+          description=None,
+          cols=Right(Nil),
+          outputView=outputView, 
+          input=Right(emptyDirectory),
+          authentication=None,
+          persist=false,
+          numPartitions=None,
+          contiguousIndex=true,
+          params=Map.empty
+        )  
+      )
+    }
+    assert(thrown1.getMessage === "BytesExtract has produced 0 columns and no schema has been provided to create an empty dataframe.")
+    
+    // try without providing column metadata
+    val thrown2 = intercept[Exception with DetailException] {
+      val extractDataset = extract.BytesExtract.extract(
+        BytesExtract(
+          name="dataset",
+          description=None,
+          cols=Right(Nil),
+          outputView=outputView, 
+          input=Right(missingDirectory),
+          authentication=None,
+          persist=false,
+          numPartitions=None,
+          contiguousIndex=true,
+          params=Map.empty
+        )  
+      )
+    }
+    assert(thrown2.getMessage === "BytesExtract has produced 0 columns and no schema has been provided to create an empty dataframe.")
+
+    // try with column
+    val extractDataset = extract.BytesExtract.extract(
+      BytesExtract(
+        name="dataset",
+        description=None,
+        cols=Right(cols),
+        outputView=outputView, 
+        input=Right(emptyWildcardDirectory),
+        authentication=None,
+        persist=false,
+        numPartitions=None,
+        contiguousIndex=true,
+        params=Map.empty
+      )  
+    ).get
+
+    assert(extractDataset.columns.deep == Array("binaryDatum", "_filename").deep)
+    assert(extractDataset.count == 0)
+  }  
 
 }